requirements.txt
# LiveKit dependencies
livekit>=0.12.0
livekit-agents==0.8.0
livekit-plugins-openai==0.6.0

# Groq
groq==0.33.0

# Firebase
firebase-admin==6.5.0

# Web framework for supervisor UI
flask==3.0.0
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
